Magic hasarları için düzenleme

+++ src/GameServer/MagicInstance.cpp
@@ -2273,7 +2273,7 @@ bool MagicInstance::ExecuteType9()
 
 short MagicInstance::GetMagicDamage(Unit *pTarget, int total_hit, int attribute)
 {	
-	int32 damage = 0, temp_hit = 0, righthand_damage = 0, attribute_damage = 0;
+	short damage = 0, temp_hit = 0, righthand_damage = 0, attribute_damage = 0;
 	int random = 0, total_r = 0 ;
 	uint8 result; 
 
@@ -2406,14 +2406,14 @@ short MagicInstance::GetMagicDamage(Unit *pTarget, int total_hit, int attribute)
 
 		}
 
-		damage = (short)(230 * total_hit / (total_r + 250));
+		damage = (230 * total_hit / (total_r + 250));
 		random = myrand(0, damage);
-		damage = (short)(random * 0.3f + (damage * 0.85f)) - sMagicAmount;
+		damage = (random * 0.3f + (damage * 0.85f)) - sMagicAmount;
 
 		if (pSkillCaster->isNPC())
 			damage -= ((3 * righthand_damage) + (3 * attribute_damage));
 		else if (attribute != MAGIC_R)	// Only if the staff has an attribute.
-			damage -= (short)(((righthand_damage * 0.8f) + (righthand_damage * pSkillCaster->GetLevel()) / 60) + ((attribute_damage * 0.8f) + (attribute_damage * pSkillCaster->GetLevel()) / 30));
+			damage -= (((righthand_damage * 0.8f) + (righthand_damage * pSkillCaster->GetLevel()) / 60) + ((attribute_damage * 0.8f) + (attribute_damage * pSkillCaster->GetLevel()) / 30));
 		if (pTarget->m_bMagicDamageReduction < 100)
 			damage = damage * pTarget->m_bMagicDamageReduction / 100;
 	}
@@ -2429,7 +2429,7 @@ short MagicInstance::GetMagicDamage(Unit *pTarget, int total_hit, int attribute)
 	if (damage > MAX_DAMAGE)
 		damage = MAX_DAMAGE;
 
-	return (short)(damage);
+	return damage;
 }
 
 int32 MagicInstance::GetWeatherDamage(int32 damage, int attribute)
